+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Svelte Router changelog
 
+## 5.2.0
+
+- Internationalisation. Translate the route names to as many languages as you need.
+
+```javascript
+const adminRoutes = [
+  {
+    name: 'employees',
+    layout: EmployeesPage,
+    lang: { es: 'empleados', fr: 'employes', de: 'angestellte', it: 'impiegati' }
+  }
+]
+```
+
 ## 5.1.1
 
 - Fix an error when a route has a redirectTo that leads to a guarded route that also redirects.
@@ -7,20 +21,21 @@
 ```javascript
 const adminRoutes = [
   {
-    name: "/admin",
+    name: '/admin',
     layout: AdminLayout,
     nestedRoutes: [
-      { name: "index", redirectTo: "admin/dashboard" },
+      { name: 'index', redirectTo: 'admin/dashboard' },
       {
-        name: "dashboard",
+        name: 'dashboard',
         component: DashboardIndex,
         onlyIf: {
           guard: isLoggedIn,
-          redirect: "/login"
+          redirect: '/login'
         }
-      },
+      }
+    ]
   }
-}
+]
 ```
 
 ## 5.1.0